smartdrive-media-extractor/pyproject.toml
[project]
name = "smartdrive-media-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "faster-whisper>=1.1.1",
    "ffmpeg>=1.4",
    "ffmpeg-python>=0.2.0",
    "flask>=3.1.1",
    "google-cloud-pubsub>=2.30.0",
    "google-cloud-storage>=3.1.1",
    "google-generativeai>=0.8.5",
    "transformers>=4.52.4",
    "weaviate>=0.1.2",
    "weaviate-client>=4.15.3",
]
